requirements.txt
alembic==1.15.1
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
spotipy==2.25.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
email_validator==2.2.0
alembic==1.15.1
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.25.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
